+++ src/Settings/Preview/vitals.lua
@@ -5,7 +5,7 @@
 local TR = _G.LUI.Locale.TR
 local lui_tokenize_format = _G.LUI.Utils.lui_tokenize_format
 local lui_format_tokenized = _G.LUI.Utils.lui_format_tokenized
-local lui_format_timeout = _G.LUI.Utils.lui_format_timeout
+local lui_format_icon_timeout = _G.LUI.Utils.lui_format_icon_timeout
 local lui_vitals_layout_label = _G.LUI.Utils.lui_vitals_layout_label
 local lui_apply_opacity_to_color = _G.LUI.Utils.lui_apply_opacity_to_color
 local lui_abbrev_number = _G.LUI.Utils.lui_abbrev_number
@@ -426,8 +426,8 @@ function StandardVitalsPreview:update()
         if t < 0 then
             return ""
         end
-        if t < 9 then
-            return lui_format_timeout(t)
+        if t < 10 then
+            return lui_format_icon_timeout(t)
         end
         return ""
     end